evaluacion-TSPLIB/requirements.txt
click==6.7
hydra-core==1.3.2
mock==3.0.5
numpy==1.24.2
ortools==9.3.10497
pytest==5.1.2
pytest-cov==2.7.1
twine==1.14.0
tsplib95==0.7.1

